+++ src/tests/fixtures/input_type_one_of/oneOfFieldMissingTypeAnnotation.invalid.ts
@@ -1,0 +1,5 @@
+/**
+ * @gqlInput
+ * @oneOf
+ */
+export type Greeting = { firstName: string } | { lastName };