+++ src/tests/fixtures/input_type_one_of/oneOfMemberHasNoField.invalid.ts
@@ -1,0 +1,5 @@
+/**
+ * @gqlInput
+ * @oneOf
+ */
+type Greeting = { firstName: string } | {};