+++ src/tests/fixtures/input_type_one_of/oneOfNotOnType.invalid.ts
@@ -1,0 +1,5 @@
+/**
+ * @gqlInput
+ * @oneOf
+ */
+export class Greeting {}