+++ src/tests/fixtures/input_type_one_of/oneOfWithFieldDescriptionInsideObj.ts
@@ -1,0 +1,15 @@
+// Known issue, descriptions are not parsed?
+
+/**
+ * @gqlInput
+ * @oneOf
+ */
+export type Greeting =
+  | {
+      /** First Name */
+      firstName: string;
+    }
+  | {
+      /** Last Name */
+      lastName: string;
+    };